feat(nebu-plugin): ensure top-down component evaluation
When a composite element is nested within another composite element (without a closure between them), this commit wraps the descendant in a thunk to ensure the parent element is evaluated first.

This means the `children` prop of a function component may contain a thunk, which means they can inherit context provided by the function component.

Note that element props are bound at the time of component evaluation. In all practical cases, this is never an issue. Just be aware that parent element props are bound first.

Lastly, element props whose value is a composite element will not have that "prop element" wrapped in a thunk. It was decided that would be too clunky for function components to handle. Better to ensure that component props are never unexpected thunks instead of Element nodes.