feat: add `AlienHooks#enableContext`
This is a shorthand for providing context while an element is mounted.

To be clear, context is provided globally. You don't have to be a descendant of the element that the context's lifecycle is tied to. So in most cases, this technique is only useful if you can be sure there won't be multiple instances of the component that creates the associated element.