fix: support adding component instance to DOM with native methods
In most cases, a component's root node will be added as a child to another JSX node, and so jsx-dom will enable the component hooks when that happens.

But if a component's root node is added to the DOM with native methods (eg: direct use of appendChild), this commit will detect that in the next microtask and tie the component hooks to a mutation observer (instead of relying on jsx-dom for enabling and morphdom for disabling).