feat: add `Transition` component
This component is useful for animating between two elements. The `id` prop identifies the `children` prop, so a transition can be gracefully interrupted during animation (rather than abruptly mounting a new element) if the `id` prop is changed to match a leaving element.

This component is also a good example of using the new <ManualUpdates> element and `updateNode` function. Leaving elements are wrapped in an absolute-positioned container, so as to not affect the layout of the current children. Normal use of alien-dom prohibits moving children of a self-updating component to a new parent element, because alien-dom uses element keys to match elements only in the same parent. This is a problem for the Transition component, but the ManualUpdates API is an easy workaround. Yet, the Transition component still wants children to be update-able from the parent component, so it calls `updateNode` to handle that.