feat: export `defineChannel` function
BREAKING CHANGE

This removes the following APIs:
1. The `dispatch` and `on` element methods, which were confusing and not strongly typed.
2. The `events` object, which was not strongly typed.
3. The `AlienElementMessage` type was replaced with `AlienBubblingMessage` to better reflect its purpose.

The dispatch & on methods have been replaced by the `defineChannel` function, which creates a single event channel whose event data is strongly typed. Event channels have "receivers" (a.k.a. event listeners), which may be global or targeted at a specific element. Similar to how the old API worked, targeted receivers will be called if an event is sent to a descendant or to the target itself. Unlike the old API, global receivers will be called (last) for targeted events, whereas they were not called at all previously.

Events sent via the channels API can have their propagation stopped just like DOM events can. For targeted events, the `stopPropagation` method works as expected (receivers targeting an ancestor won't be called; see MDN docs for more info). For untargeted events, it does nothing while `stopImmediatePropagation` *will* work normally.

To send an event through an event channel, simply call the event channel like a function (passing event data with it). To receive an event, pass a function as the last argument:

    const onThing = defineChannel()
    onThing(() => console.log('thing happened'))
    onThing() // → Logs the above message.

To send a targeted event, pass the element as the first argument.

⚠️ An event channel's custom data must be a plain object type or nothing.

Last but not least, you can use array destructuring on an event channel to split it into two distinct sendMessage and addReceiver functions:

    export const [doThing, onThing] = defineChannel()
    onThing(() => console.log('thing happened'))
    doThing()

This is useful when exporting the channel for other modules to trigger an event with, as it lets you use clear naming for the "send message" use case. In other words, calling `doThing` feels more like an action, while calling `onThing` feels like a reaction.