feat: support nested refs in `style` prop
Similar to the `class` prop, the observable refs can be nested in the `style.width` property for example, whose object might even be nested in an array which is then passed as the `style` prop.

This commit also adds flattening of the style array before it's used to update the element style. This helps in avoiding unnecessary updates due to multiple refs under the same key (but defined in separate style objects).