refactor(observable): take a new approach with ComputedRef
I have removed the `ReadonlyRef::_version` property, because it is not reliable for change detection where synchronous updates are allowed. Also, instead of using `emptySymbol` to determine the dirty state of a ComputedRef, I have added an explicit dirty flag. This allows a ComputedRef to retain its old value for comparison purposes. Another new addition is the idea of a temporary observer in the context of a ComputedRef. This is responsible for marking a ComputedRef dirty (in a synchronous manner) when a persistent observer is not being used. This ensures the ComputedRef always returns an up-to-date value.

A few other changes:
- Added the `Observer::_access` and `Observer::_update` methods for logic reuse
- The `__OBSERVABLE_HOOKS__.didUpdate` function is now wrapped with `peek` to avoid accidental observation
- If a ComputedRef is accessed synchronously, the asynchronous update is now a no-op (it avoids unnecessary recompute)
- Ref access within an `Observer::onUpdate` callback is now consistently observed (so you should use `peek` if observation is undesirable)
- There is now an `Observer` namespace with callback types and a queue type
- Speaking of observer queues, the queue argument of `new Observer` is now looser (no Set required)