feat: deferred component updates
Instead of `jsx()` creating a DOM node every time, a "deferred node" can be returned. This happens when a keyed JSX element has an existing DOM node and when an JSX element has no defined key.

A deferred node is effectively a VDOM node. This allows for unkeyed elements to still be used to patch compatible DOM nodes. By allowing for this, unkeyed JSX elements declared in a loop or render function will not be remounted if a compatible node is found.

As always, "element references" are not permitted in loops and render functions unles the `key` prop is defined, so it is safe for `jsx()` to return a deferred node when no key is defined.