feat(machine): rework the MachineProxy type and Machine::assert method
The assert method now returns the state, instead of invoking a callback. State should be destructured as needed. Methods should not be reliant on "this".

The proxy type now lets you access any state. Invalid access is enforced at runtime, instead of trying to achieve static type safety (which is not easy with the way defineMachine is designed).