fix(compiler): clear auto-memoized object when observed ref is changed
If an auto-memoized object/array contains a function call, we have to wrap it in a thunk and collect any observable refs that were possibly accessed by the function call. Then any time a component observer is notified of a changed ref, we check the refs we collected earlier to see if any of the auto-memoized objects/arrays need to be cleared.