fix(useObserver): update `onChange` callback transparently
When a ref is passed as the first argument, the onChange callback can be updated transparently (i.e. without a deps array), since the references within the callback are never observed.

Apart from a better DX (not having to manually define a deps array), the other motivation was to fix a bug where the onChange callback was being called unnecessarily when the deps changed.