chore: remove `selfUpdating` and `markPureComponent` functions
"Pure components" were never fully supported nor was their behavior clearly specified. Internally, there is still the concept of a stateless component (which skips the AlienComponent instance), but this is only used by context providers and shadow roots.

The `selfUpdating` function was an internal remnant of a humbler past, so there‘s no point in continuing to wrap components with it when it can just be baked into the JSX runtime (obviating the need for a wrapper component).