fix(animate): remember non-null `from` prop until it‘s applied
If one `animate` call sets the `from` prop and another `animate` call sets the `to` prop, use the `from` prop from the first call unless it‘s overridden. This can happen when a delay is desired but the initial value should be applied synchronously. To achieve that, you have to make two calls to `animate` (one with `from` set and another with `to` and `delay` set).