feat(BREAKING): clean up TypeScript types
Biggest change here is that support for `ReadonlyRef` in attributes is now exclusively in the `JSX` namespace. So the `HTMLClassAttribute` type does not allow observable refs, while `JSX.HTMLClassProp` does. A few other examples include `JSX.HTMLStyleProp`, `JSX.CSSProps`, and `JSX.HTMLDatasetProp`.

Also, the `key`, `ref`, and `children` attributes are *only* defined in the `JSX` namespace, as these attributes don't exist in the context of traditional HTML/SVG elements (which is what HTMLAttributes and SVGAttributes represent).

BREAKING CHANGE