chore(refactor): stop relying on MutationObserver for mount detection
Use microtask polling of `node.isConnected` instead, since it doesn‘t inadvertently produce a bunch of unnecessary work for the garbage collector in the form of unrelated mutation events.

This also allows the logic in "addons/domObserver.ts" to be tree-shaked. To achieve this, I also had to separate the `AlienEffects#enableOnceMounted` method into its own subclass "AlienMountEffects" which is now only used by the `AlienElement#effects()` extension method.