chore(feat): allow for arbitrary NodeStore contexts
Now any object can implement the NodeStore interface to act as a cache for DOM nodes and their deferred updates. This will be most useful for the ArrayView feature, where JSX elements can be rendered outside of a component context but they still need to be cached somewhere for updating purposes.