feat: add useArrayView hook and ArrayView component
“Array views” let you render an array on a granular level, so you don‘t need to rerender every array item just because the array changed in some little way. This takes advantage of the new “array observer” feature.

To avoid re-rendering every array item when the parent component re-renders, you should provide a deps array to the useArrayView hook. This shouldn‘t be necessary with the ArrayView component, since the ‘render prop’ should be memoized automatically by the AlienDOM compiler, but you can be explicit with the `deps` prop if you need more control.