chore(refactor): better code readability + remove `arguments` access in hot path
In the `createSymbolProperty` factory function, a function was returned that could be used as either a getter or a setter. To accomplish this, it checked the `arguments.length` which is not great for performance, especially considering how much the private symbols are used throughout AlienDOM's hot paths.

Additionally, seeing a bunch of odd `kAlienFooBar(obj)` and `kAlienFooBar(obj, value)` calls everywhere is not the most welcoming style for new readers of the codebase.